refactoring editor header
