+++ app/client/src/pages/Editor/EditorName/types.ts
@@ -1,0 +1,6 @@
+export enum MenuTypes {
+  MENU = "menu",
+  PARENT = "parent",
+  RECONFIRM = "re-confirm",
+  MENU_DIVIDER = "menu divider",
+}
